refactor: move MemgraphIngestor into services/graph behind a shim
